requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python-headless
pillow
python-multipart
numpy
